[0301]PH

+++ app/controllers/company/employments_controller.rb
@@ -9,18 +9,21 @@ class Company::EmploymentsController < ApplicationController
     @employment.company_id = current_company.id
     @employment.staff_id = current_company.request_forms.first.staff_id
     @employment.employment_status = 0
-    @employment.save
 
-    current_company.request_forms.each do |request_form|
-      @employment_details = EmploymentDetail.new
-      @employment_details.employment_id = @employment.id
-      @employment_details.request_id = request_form.request.id
-      @employment_details.total_payment_amount = @employment.total_payment_amount
-      @employment_details.save
-      end
+    if  @employment.save
+        current_company.request_forms.each do |request_form|
+          @employment_details = EmploymentDetail.new
+          @employment_details.employment_id = @employment.id
+          @employment_details.request_id = request_form.request.id
+          @employment_details.total_payment_amount = @employment.total_payment_amount
+          @employment_details.save
+          end
 
       current_company.request_forms.destroy_all
       redirect_to  company_employments_complete_path
+    else
+      render action: :new
+    end
   end
 
   def confirm
@@ -35,8 +38,12 @@ class Company::EmploymentsController < ApplicationController
 
   def index
     @employments = Employment.where(company_id: current_company.id)
+    if @employments.empty?
+    @employments = []
+    else
     @employment_details = EmploymentDetail.find_by(employment_id: @employments.first.id)
     @request_id = @employment_details.request_id
+    end
   end
 
   def show
@@ -48,7 +55,7 @@ class Company::EmploymentsController < ApplicationController
   private
 
   def employment_params
-  params.require(:employment).permit(:company_id, :staff_id, :name, :kana, :response_deadline, :date,  :start_time,  :finish_time, :hours, :hourly_wage, :payment_method, :place_of_employment, :introduction, :contact_address, :total_payment_amount, :employment_status)
+  params.require(:employment).permit(:company_id, :staff_id, :employment_name, :employment_kana, :response_deadline, :date,  :start_time,  :finish_time, :hours, :hourly_wage, :payment_method, :place_of_employment, :introduction, :contact_address, :total_payment_amount, :employment_status)
   end
