+++ app/controllers/staff/works_controller.rb
@@ -9,18 +9,20 @@ class Staff::WorksController < ApplicationController
     @work.staff_id = current_staff.id
     @work.company_id = current_staff.recruitment_forms.first.company_id
     @work.work_status = 0
-    @work.save
 
-    current_staff.recruitment_forms.each do |recruitment_form|
-      @work_details = WorkDetail.new
-      @work_details.work_id = @work.id
-      @work_details.recruitment_id = recruitment_form.recruitment.id
-      @work_details.total_payment_amount = @work.total_payment_amount
-      @work_details.save
-    end
-
-    current_staff.recruitment_forms.destroy_all
-    redirect_to  staff_works_complete_path
+    if  @work.save
+        current_staff.recruitment_forms.each do |recruitment_form|
+          @work_details = WorkDetail.new
+          @work_details.work_id = @work.id
+          @work_details.recruitment_id = recruitment_form.recruitment.id
+          @work_details.total_payment_amount = @work.total_payment_amount
+          @work_details.save
+        end
+        current_staff.recruitment_forms.destroy_all
+        redirect_to  staff_works_complete_path
+      else
+        render action: :new
+      end
   end
 
   def confirm
@@ -36,8 +38,12 @@ class Staff::WorksController < ApplicationController
 
   def index
     @works = Work.where(staff_id: current_staff.id)
-    @work_details = WorkDetail.find_by(work_id: @works.first.id)
-    @recruitment_id = @work_details.recruitment_id
+    if @works.empty?
+      @works = []
+    else
+      @work_details = WorkDetail.find_by(work_id: @works.first.id)
+      @recruitment_id = @work_details.recruitment_id
+    end
   end
 
   def show
@@ -49,7 +55,7 @@ class Staff::WorksController < ApplicationController
   private
 
   def work_params
-    params.require(:work).permit(:staff_id, :company_id, :name, :kana, :response_deadline, :date,  :start_time,  :finish_time, :hours, :transportation, :payment_method, :whereabouts, :introduction, :contact_address, :total_payment_amount, :work_status)
+    params.require(:work).permit(:staff_id, :company_id, :work_name, :work_kana, :response_deadline, :date,  :start_time,  :finish_time, :hours, :transportation, :payment_method, :whereabouts, :introduction, :contact_address, :total_payment_amount, :work_status)
   end
 
 end
